refactor(customer-host): moves App.vue to presentation layer
