+++ customer/src/main.ts
@@ -1,7 +1,7 @@
 import { createApp } from 'vue'
 import { createPinia } from 'pinia'
 import piniaPluginPersistedstate from 'pinia-plugin-persistedstate'
-import App from './App.vue'
+import App from './presentation/views/App.vue'
 
 const pinia = createPinia()
 const app = createApp(App)
